+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "auth-astro",
-  "version": "1.0.1",
+  "version": "1.0.2",
   "main": "./src/index.ts",
   "exports": {
     ".": {
@@ -10,10 +10,6 @@
     "./client": {
       "types": "./src/client.ts",
       "import": "./src/client.ts"
-    },
-    "./server": {
-      "types": "./src/server.ts",
-      "import": "./src/server.ts"
     },
     "./components": {
       "types": "./src/components.ts",
@@ -36,7 +32,7 @@
   "devDependencies": {
     "@types/cookie": "^0.5.1",
     "@types/node": "^18.11.18",
-    "astro": "^2.0.0",
+    "astro": "^2.0.2",
     "next-auth": "^4.18.8",
     "typescript": "^4.9.4"
   },
@@ -44,5 +40,14 @@
     "@auth/core": "^0.2.5",
     "set-cookie-parser": "^2.5.1"
   },
-  "keywords": ["auth", "astro", "authentication", "integration", "@auth", "authjs", "jwt", "oauth"]
+  "keywords": [
+    "auth",
+    "astro",
+    "authentication",
+    "integration",
+    "@auth",
+    "authjs",
+    "jwt",
+    "oauth"
+  ]
 }
