+++ src/components/Auth.astro
@@ -1,5 +1,5 @@
 ---
-import { type AstroAuthConfig, getSession } from '../server'
+import { type AstroAuthConfig, getSession } from '..'
 
 interface Props {
   authOpts: AstroAuthConfig
