+++ src/components/Signin.astro
@@ -11,7 +11,6 @@ interface Props extends HTMLAttributes<'button'> {
 }
 
 const { provider, options, authParams, ...attrs } = Astro.props
-
 attrs.class = `sign-in ${attrs.class ?? ''}`
 ---
